first phase refactor. Missing load function in formatters for enabling formatters discovery [skip-ci]
